UPDATE ::: Adding null checks
git-svn-id: svn://107.170.10.188:8080/Monolith/dev@26912 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 76e17b4920ab4c477521f06b1bc1d71736f8ad06

+++ src/prerna/upload/ImageUploader.java
@@ -50,8 +50,10 @@ public class ImageUploader extends Uploader {
 		// and delete them
 		File[] oldImages = findImageFile(f.getParentFile());
 		// delete if any exist
-		for(File oldI : oldImages) {
-			oldI.delete();
+		if(oldImages != null) {
+			for(File oldI : oldImages) {
+				oldI.delete();
+			}
 		}
 		writeFile(imageFile, f);
 	}
